1. 修正SamFile的bug，关闭后可以继续随机读写

+++ src/main/java/com/novelbio/analysis/seq/sam/SamReader.java
@@ -175,7 +175,7 @@ public class SamReader {
 		if (readByte < 0 || filesize < 0) {
 			return -1;
 		} else {
-			return readByte/filesize;
+			return (double)readByte/filesize;
 		}
 	}
